test(qtensor): remove duplicated test
