+++ test/test_serialization.py
@@ -4,32 +4,8 @@ import pytest
 import torch
 from helpers import random_qtensor, random_tensor
 
-from quanto import Calibration, QTensor, absmax_scale, freeze, qfloat8, qint8, quantize
+from quanto import Calibration, freeze, qint8, quantize
 from quanto.nn import QLinear, QModuleMixin
-
-
-@pytest.mark.parametrize("input_shape", [(10,), (1, 10), (2, 10), (10, 32, 32)])
-@pytest.mark.parametrize("qtype", [qint8, qfloat8], ids=["qint8", "qfloat8"])
-@pytest.mark.parametrize("dtype", [torch.float16, torch.float32], ids=["fp16", "fp32"])
-@pytest.mark.parametrize("axis", [None, 0, -1], ids=["per-tensor", "first-axis", "last-axis"])
-def test_quantized_tensor_serialization(input_shape, qtype, dtype, axis):
-    inputs = random_tensor(input_shape, dtype=dtype)
-    scale = absmax_scale(inputs, qtype, axis)
-    qinputs = QTensor.quantize(inputs, qtype, scale)
-    b = io.BytesIO()
-    torch.save(qinputs, b)
-    b.seek(0)
-    qinputs_reloaded = torch.load(b)
-    assert qinputs_reloaded.qtype == qtype
-    assert torch.equal(qinputs_reloaded._scale, qinputs._scale)
-    if qtype.is_floating_point:
-        # Equality is not supported for float8
-        assert torch.equal(qinputs_reloaded._data.to(torch.float32), qinputs._data.to(torch.float32))
-    else:
-        assert torch.equal(qinputs_reloaded._data, qinputs._data)
-    # We cannot test dtype directly as it is not correctly set by torch.load
-    assert qinputs_reloaded._scale.dtype == dtype
-    assert qinputs_reloaded.axis == qinputs.axis
 
 
 @pytest.mark.parametrize("use_bias", [True, False], ids=["bias", "no-bias"])
